services/gateway_service/pyproject.toml
[tool.poetry]
name = "gateway-service"
version = "0.1.0"
description = "API Gateway service for my-scaffold-project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
grpcio = "^1.72.1"
grpcio-tools = "^1.72.1"
pydantic = {extras = ["email"], version = "^2.11.5"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
pytest-mock = "^3.12.0"
grpcio-testing = "^1.72.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
    "-v"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "app/__pycache__/*",
    "tests/*",
    "*/generated_contracts/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 